Move all the logic from SmartLock to AbstractSmartLock
